Fix protocol error

+++ src/main/java/mpo/dayon/common/network/message/NetworkResizeScreenMessage.java
@@ -44,7 +44,7 @@ public class NetworkResizeScreenMessage extends NetworkMessage {
 
 	public static NetworkResizeScreenMessage unmarshall(ObjectInputStream in) throws IOException {
 		final int width = in.readInt();
-		final int height = in.readShort();
+		final int height = in.readInt();
 		return new NetworkResizeScreenMessage(width, height);
 	}
